fixing generated types with proxy modules
