+++ packages/destination-actions/src/destinations/engage/sendgrid/generated-types.ts
@@ -1,32 +1,1 @@
-// Generated file. DO NOT MODIFY IT BY HAND.
-
-export interface Settings {
-  /**
-   * The API key for your Unlayer account
-   */
-  unlayerApiKey?: string
-  /**
-   * The Api Key for your SendGrid account
-   */
-  sendGridApiKey: string
-  /**
-   * Profile API Environment
-   */
-  profileApiEnvironment: string
-  /**
-   * Profile API Access Token
-   */
-  profileApiAccessToken: string
-  /**
-   * Space ID
-   */
-  spaceId: string
-  /**
-   * Source ID
-   */
-  sourceId: string
-  /**
-   * The region where the email is originating from
-   */
-  region?: string
-}
+export * from '../../engage-messaging-sendgrid/generated-types'
